requirements.txt
flask==3.0.3
flask-cors==4.0.1
duckdb==0.10.3
groq==0.13.0
openai==1.35.13
python-dotenv==1.0.1
pandas==2.2.2
pyarrow==16.1.0
sendgrid==6.11.0
